updated test scripts
