+++ Testing/k6/src/application/load/docker-compose.yml
@@ -2,9 +2,9 @@ version: '3.4'
 services:
   k6:
     image: loadimpact/k6:latest
-    command: run /services/profile/k6_profile.js
+    command: run /application/load/k6_load.js
     volumes:
-      - ../../../src/services/profile:/services/profile
+      - ../../../src/application/load:/application/load
       - ../../../src/inc:/inc
     environment: 
       - "HG_TYPE=${HG_TYPE}"
