+++ Testing/k6/src/application/smoke/docker-compose.yml
@@ -2,12 +2,13 @@ version: '3.4'
 services:
   k6:
     image: loadimpact/k6:latest
-    command: run /services/profile/k6_profile.js
+    command: run /application/smoke/k6_smoke.js
     volumes:
-      - ../../../src/services/profile:/services/profile
+      - ../../../src/application/smoke:/application/smoke
       - ../../../src/inc:/inc
     environment: 
       - "HG_TYPE=${HG_TYPE}"
       - "HG_ENV=${HG_ENV}"
       - "HG_KEY=${HG_KEY}"
       - "HG_PASSWORD=${HG_PASSWORD}"
+
