+++ Testing/k6/src/application/smoke/k6_smoke.js
@@ -13,30 +13,28 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 //-------------------------------------------------------------------------
-
 import http from "k6/http";
-import { sleep } from "k6";
-import * as common from "./inc/common.js";
+import { check, group, sleep } from "k6";
+import { Rate, Trend } from "k6/metrics";
+import * as common from "../../inc/common.js";
 
-export let options = {
-    stages: [
-        { duration: "1m", target: 10 }, // below normal load
-        { duration: "2m", target: 250 },
-        { duration: "3h56m", target: 250 }, // stay at high users for hours 'soaking' the system
-        { duration: "2m", target: 0 }, // drop back down
-    ],
-};
+export let options = common.smokeOptions;
 
 export default function () {
-    let user = common.users[__VU % common.users.length];
+    let user =
+        common.users[common.getRandomInteger(0, common.users.length - 1)];
 
     common.authorizeUser(user);
 
-    let webClientBatchResponses = http.batch(common.webClientRequests(user));
-    let timelineBatchResponses = http.batch(common.timelineRequests(user));
+    common.groupWithDurationMetric("batch", function () {
+        let webClientBatchResponses = http.batch(
+            common.webClientRequests(user)
+        );
+        let timelineBatchResponses = http.batch(common.timelineRequests(user));
 
-    common.checkResponses(webClientBatchResponses);
-    common.checkResponses(timelineBatchResponses);
+        common.checkResponses(webClientBatchResponses);
+        common.checkResponses(timelineBatchResponses);
+    });
 
-    sleep(common.getRandom(1.0, 3.0));
+    sleep(1);
 }
