+++ Testing/k6/src/application/soak/k6_soak.js
@@ -16,18 +16,27 @@
 
 import http from "k6/http";
 import { sleep } from "k6";
-import * as common from "../inc/common.js";
+import * as common from "../../inc/common.js";
 
-export let options = common.OptionConfig();
+export let options = {
+    stages: [
+        { duration: "1m", target: 10 }, // below normal load
+        { duration: "2m", target: 250 },
+        { duration: "3h56m", target: 250 }, // stay at high users for hours 'soaking' the system
+        { duration: "2m", target: 0 }, // drop back down
+    ],
+};
 
 export default function () {
     let user = common.users[__VU % common.users.length];
 
     common.authorizeUser(user);
-    let response = http.get(
-        common.ConfigurationUrl + "/" + user.hdid,
-        common.params(user)
-    );
-    common.checkResponse(response);
-    sleep(1);
+
+    let webClientBatchResponses = http.batch(common.webClientRequests(user));
+    let timelineBatchResponses = http.batch(common.timelineRequests(user));
+
+    common.checkResponses(webClientBatchResponses);
+    common.checkResponses(timelineBatchResponses);
+
+    sleep(common.getRandom(1.0, 3.0));
 }
