+++ Testing/k6/src/application/stress/docker-compose.yml
@@ -2,9 +2,9 @@ version: '3.4'
 services:
   k6:
     image: loadimpact/k6:latest
-    command: run /services/profile/k6_profile.js
+    command: run /application/stress/k6_stress.js
     volumes:
-      - ../../../src/services/profile:/services/profile
+      - ../../../src/application/stress:/application/stress
       - ../../../src/inc:/inc
     environment: 
       - "HG_TYPE=${HG_TYPE}"
