+++ Testing/k6/src/application/stress/k6_stress.js
@@ -16,24 +16,32 @@
 
 import http from "k6/http";
 import { sleep } from "k6";
-import * as common from "./inc/common.js";
+import * as common from "../../inc/common.js";
 
-export let options = common.loadOptions;
+export let options = {
+    stages: [
+        { duration: "2m", target: 50 }, // below normal load
+        { duration: "5m", target: 100 },
+        { duration: "2m", target: 200 }, // normal load
+        { duration: "5m", target: 200 },
+        { duration: "2m", target: 400 }, // around the breaking point
+        { duration: "4m", target: 400 },
+        { duration: "2m", target: 500 }, // beyond the breaking point
+        { duration: "5m", target: 550 },
+        { duration: "5m", target: 0 }, // scale down. Recovery stage.
+    ],
+};
 
 export default function () {
     let user = common.users[__VU % common.users.length];
 
     common.authorizeUser(user);
 
-    common.groupWithDurationMetric("batch", function () {
-        let webClientBatchResponses = http.batch(
-            common.webClientRequests(user)
-        );
-        let timelineBatchResponses = http.batch(common.timelineRequests(user));
+    let webClientBatchResponses = http.batch(common.webClientRequests(user));
+    let timelineBatchResponses = http.batch(common.timelineRequests(user));
 
-        common.checkResponses(webClientBatchResponses);
-        common.checkResponses(timelineBatchResponses);
-    });
+    common.checkResponses(webClientBatchResponses);
+    common.checkResponses(timelineBatchResponses);
 
-    sleep(1);
+    sleep(common.getRandom(0.5, 3.0));
 }
