+++ Testing/k6/src/services/patient/docker-compose.yml
@@ -10,3 +10,4 @@ services:
       - "HG_TYPE=${HG_TYPE}"
       - "HG_ENV=${HG_ENV}"
       - "HG_KEY=${HG_KEY}"
+      - "HG_PASSWORD=${HG_PASSWORD}"
