+++ Testing/k6/src/webapp_api/communication/docker-compose.yml
@@ -2,9 +2,9 @@ version: '3.4'
 services:
   k6:
     image: loadimpact/k6:latest
-    command: run /services/profile/k6_profile.js
+    command: run /webapp_api/communication/k6_communication.js
     volumes:
-      - ../../../src/services/profile:/services/profile
+      - ../../../src/webapp_api/communication:/webapp_api/communication
       - ../../../src/inc:/inc
     environment: 
       - "HG_TYPE=${HG_TYPE}"
