+++ Testing/k6/src/webapp_api/configuration/k6_configuration.js
@@ -16,32 +16,18 @@
 
 import http from "k6/http";
 import { sleep } from "k6";
-import * as common from "./inc/common.js";
+import * as common from "../../inc/common.js";
 
-export let options = {
-    stages: [
-        { duration: "2m", target: 50 }, // below normal load
-        { duration: "5m", target: 100 },
-        { duration: "2m", target: 200 }, // normal load
-        { duration: "5m", target: 200 },
-        { duration: "2m", target: 400 }, // around the breaking point
-        { duration: "4m", target: 400 },
-        { duration: "2m", target: 500 }, // beyond the breaking point
-        { duration: "5m", target: 550 },
-        { duration: "5m", target: 0 }, // scale down. Recovery stage.
-    ],
-};
+export let options = common.OptionConfig();
 
 export default function () {
     let user = common.users[__VU % common.users.length];
 
     common.authorizeUser(user);
-
-    let webClientBatchResponses = http.batch(common.webClientRequests(user));
-    let timelineBatchResponses = http.batch(common.timelineRequests(user));
-
-    common.checkResponses(webClientBatchResponses);
-    common.checkResponses(timelineBatchResponses);
-
-    sleep(common.getRandom(0.5, 3.0));
+    let response = http.get(
+        common.ConfigurationUrl + "/" + user.hdid,
+        common.params(user)
+    );
+    common.checkResponse(response);
+    sleep(1);
 }
